+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 * Add `Collection#find_by` for easier finding: `album.songs.find_by(id: 1)`.
 * Fix inheritance of the `:default` option. This would formerly wrap the default value into another `Uber::Options::Value`.
+* Introduce `Struct#save!`.
 
 # 0.2.2